AddRecordsSeedScript/requirements.txt
certifi==2020.6.20
chardet==3.0.4
Faker==4.14.0
idna==2.10
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
text-unidecode==1.3
urllib3==1.25.11
